+++ sEvolution/sGene.h
@@ -77,7 +77,7 @@ public:
 
 	void random(float limit = 1.f)
 	{
-		m_data = sRandom::getInt(0, ((1 << m_bits) - 1) * limit);
+		m_data = (unsigned int)sRandom::getInt(0, (int)(((1 << m_bits) - 1) * limit));
 	}
 
 	void invert()
